+++ src/output/shapes/passthrough.ts
@@ -52,6 +52,7 @@ const PASSTHROUGH_COMMANDS = [
   "init",
   "lint",
   "list",
+  "move",
   "proposal-accept-batch",
   "proposal-drain",
   "proposal-reject-batch",